refactor(dashboard): extract registry/agent_graph/build_lock to registry.rs (#1266)
routes.rs: 2613 -> 2517
registry.rs: NEW 105

Moves registry_list/register/deregister/reap, agent_graph,
build_lock_status/force_release handlers.
